+++ src/main/java/com/team2502/robot2017/command/FlywheelCommand.java
@@ -41,7 +41,7 @@ public class FlywheelCommand extends Command
         shooterSubsystem.isFlywheelActive = false;
         shooterSubsystem.isFeederActive = false;
         
-        shooterSubsystem.defaultShooter();
+//        shooterSubsystem.defaultShooter();
     }
 
     @Override
