+++ src/main/java/com/team2502/robot2017/command/LeverCommand.java
@@ -2,7 +2,7 @@ package com.team2502.robot2017.command;
 
 import com.team2502.robot2017.Robot;
 import com.team2502.robot2017.subsystem.DriveTrainTransmissionSubsystem;
-import com.team2502.robot2017.subsystem.GearSubsystem;
+import com.team2502.robot2017.subsystem.GearBoxSubsystem;
 import edu.wpi.first.wpilibj.command.Command;
 
 public class LeverCommand extends Command
