+++ src/main/java/com/team2502/robot2017/command/autonomous/GearCommandG.java
@@ -9,7 +9,7 @@ public class GearCommandG extends CommandGroup
 {
     public GearCommandG(boolean on)
     {
-        addSequential(new LeverCommand(on));
-        addSequential(new GearCommand(on));
+        //addSequential(new LeverCommand(on));
+        //addSequential(new GearCommand(on));
     }
 }
