+++ src/main/java/com/team2502/robot2017/subsystem/ShooterSubsystem.java
@@ -81,11 +81,6 @@ public class ShooterSubsystem extends Subsystem
 
         return error;
     }
-    
-    public void defaultShooter()
-    {
-    	shooterMode = false;
-    }
 	
 	public void flywheelDrive()
 	{	
@@ -104,27 +99,27 @@ public class ShooterSubsystem extends Subsystem
      	else { flywheelTalon.set(0); }
 
         // For changing the flywheel speed.
-        if(OI.JOYSTICK_DRIVE_LEFT.getRawButton(RobotMap.Joystick.Button.SHOOTER_INCREASE_SPEED))
+        if(OI.JOYSTICK_DRIVE_LEFT.getRawButton(3))
         {
             targetSpeed += 10;
         }
-        else if(OI.JOYSTICK_DRIVE_LEFT.getRawButton(RobotMap.Joystick.Button.SHOOTER_DECREASE_SPEED))
+        else if(OI.JOYSTICK_DRIVE_LEFT.getRawButton(2))
         {
             targetSpeed -= 10;
         }
 
 
         //Control for turning on/off the feeding mechanism.
-		if(OI.JOYSTICK_FUNCTION.getTrigger() && (Math.abs(flywheelTalon.getEncVelocity()) > 1650))
+		if(OI.JOYSTICK_FUNCTION.getTrigger() /*&& (Math.abs(flywheelTalon.getEncVelocity()) > Math.abs(targetSpeed - 500))*/)
 		{	
 		    feederTalon0.set(1);
 		    feederTalon1.set(-1);
 		}
-		else if(OI.JOYSTICK_FUNCTION.getTrigger() && !(Math.abs(flywheelTalon.getEncVelocity()) > 1650))
-		{
-			feederTalon0.set(1);
-			feederTalon1.set(0);
-		}
+//		else if(OI.JOYSTICK_FUNCTION.getTrigger() /*&& (Math.abs(flywheelTalon.getEncVelocity()) < Math.abs(targetSpeed - 500))*/)
+//		{
+//			feederTalon0.set(1);
+//			feederTalon1.set(0);
+//		}
 		else
 		{
 			feederTalon0.set(0);
