+++ version_master.properties
@@ -1,1 +1,1 @@
-version=0.0.3
+version=0.0.4
